feat: add docs
