+++ contracts/cosmwasm-vm/cw-mock-ibc-dapp/src/send_message.rs
@@ -25,31 +25,16 @@ impl<'a> CwIbcConnection<'a> {
             data: message,
         };
 
-        #[cfg(feature = "native_ibc")]
-        {
-            let packet = self.create_request_packet(deps, env, timeout_height, msg.clone())?;
+        let ibc_height = Height::new(0, timeout_height).unwrap();
+        let packet_data = self.create_packet(ibc_config, ibc_height, next_sequence_send, msg);
 
-            let submessage: SubMsg<Empty> =
-                SubMsg::reply_always(CosmosMsg::Ibc(packet), HOST_FORWARD_REPLY_ID);
+        println!("{} Raw Packet Created {:?}", LOG_PREFIX, &packet_data);
 
-            Ok(Response::new()
-                .add_submessage(submessage)
-                .add_attribute("method", "send_message"))
-        }
-
-        #[cfg(not(feature = "native_ibc"))]
-        {
-            let ibc_height = Height::new(0, timeout_height).unwrap();
-            let packet_data = self.create_packet(ibc_config, ibc_height, next_sequence_send, msg);
-
-            println!("{} Raw Packet Created {:?}", LOG_PREFIX, &packet_data);
-
-            let submessage: cosmwasm_std::SubMsg =
-                self.call_host_send_message(deps, info, packet_data)?;
-            Ok(Response::new()
-                .add_submessage(submessage)
-                .add_attribute("method", "send_message"))
-        }
+        let submessage: cosmwasm_std::SubMsg =
+            self.call_host_send_message(deps, info, packet_data)?;
+        Ok(Response::new()
+            .add_submessage(submessage)
+            .add_attribute("method", "send_message"))
     }
 
     fn write_acknowledgement(
@@ -61,50 +46,5 @@ impl<'a> CwIbcConnection<'a> {
     ) -> Result<Response, ContractError> {
         let _channel_id = config.src_endpoint().channel_id.clone();
         Ok(Response::new())
-    }
-}
-
-#[cfg(feature = "native_ibc")]
-impl<'a> CwIbcConnection<'a> {
-    /// This function creates an IBC message to send a packet with a timeout to a destination endpoint.
-    ///
-    /// Arguments:
-    ///
-    /// * `deps`: `deps` is a mutable reference to the dependencies of the contract. It is used to
-    /// interact with the storage and other modules of the contract.
-    /// * `env`: `env` is an object that contains information about the current blockchain environment,
-    /// such as the current block height, time, and chain ID. It is used to calculate the timeout for the
-    /// IBC packet.
-    /// * `time_out_height`: The height of the block at which the timeout for the packet will occur.
-    /// * `message`: `message` is a `CallServiceMessage` struct that contains the information needed to
-    /// create a request packet to be sent over the IBC channel. This includes the method name, input
-    /// arguments, and any other relevant data needed for the service call.
-    ///
-    /// Returns:
-    ///
-    /// a `Result` with an `IbcMsg` on success or a `ContractError` on failure.
-    fn create_request_packet(
-        &self,
-        deps: DepsMut,
-        env: Env,
-        time_out_height: u64,
-        message: Message,
-    ) -> Result<IbcMsg, ContractError> {
-        let ibc_config = self
-            .ibc_config()
-            .load(deps.as_ref().storage)
-            .map_err(ContractError::Std)?;
-
-        let timeout_block = IbcTimeoutBlock {
-            revision: 0,
-            height: time_out_height,
-        };
-        let timeout = IbcTimeout::with_both(timeout_block, env.block.time.plus_seconds(300));
-
-        Ok(IbcMsg::SendPacket {
-            channel_id: ibc_config.dst_endpoint().channel_id.clone(),
-            data: to_binary(&message).unwrap(),
-            timeout,
-        })
     }
 }
