+++ contracts/cosmwasm-vm/cw-mock-ibc-dapp/src/types/mod.rs
@@ -2,7 +2,6 @@ pub mod storage_keys;
 pub use common::rlp;
 
 pub mod config;
-pub mod dapp_msg;
 pub mod message;
 
 pub const LOG_PREFIX: &str = "[mock_ibc_dapp]:";
